+++ src/entities/task/curriculum/javascript/solutions/7_arrays_sort/6_RearrangeArray.js
@@ -1,9 +1,9 @@
-function rearrangeArray(arr) {
+const rearrangeArray = (arr) => {
   const evens = arr.filter((x) => x % 2 === 0).sort((a, b) => a - b);
   const odds = arr.filter((x) => x % 2 !== 0).sort((a, b) => a - b);
 
   return [...evens, ...odds];
-}
+};
 
 console.log(rearrangeArray([5, 3, 2, 8, 1, 4]));
 // [2, 4, 8, 1, 3, 5]