+++ src/entities/task/curriculum/javascript/solutions/8_arrays_reduce/1_Syntax.js
@@ -1,8 +1,3 @@
-const numbers = [1, 2, 3];
-
-const result = numbers.reduce((acc, curr) => {
-  return acc + curr;
-}, 0);
-
-// Пример вызова:
-console.log(result); // 6
+arr.reduce((acc, item) => {
+  return acc;
+}, initialValue);