+++ src/pages/home/ui/HomeKpiSummary.tsx
@@ -1,5 +1,5 @@
 import React, { memo } from "react";
-import { ProgressKpiGrid } from "@/entities/progress";
+import { KpiGrid } from "@/shared/ui";
 import { HomeStats } from "../model/use-home-stats";
 
 interface HomeKpiSummaryProps {
@@ -8,7 +8,7 @@ interface HomeKpiSummaryProps {
 
 export const HomeKpiSummary = memo(({ stats }: HomeKpiSummaryProps): React.JSX.Element => {
   return (
-    <ProgressKpiGrid
+    <KpiGrid
       total={stats.grandTotal}
       solved={stats.grandSolved}
       percent={stats.grandPct}